revocation: Move `crl` unit tests to their respective submodules.
